pyproject.toml
[project]
name = "capsule-learn"
version = "0.1.0"
description = "AI-powered learning assistant for TCM studies"
readme = "README.md"
authors = [
  {name = "TCM Knowledge Base", email = "bmad@example.com"}
]
maintainers = [
  {name = "TCM Knowledge Base", email = "bmad@example.com"}
]
classifiers = [
    # TODO
]
license = {text = "MIT"}
dependencies = [
  "typer",
  "python-frontmatter",
  "click>=8.0.0",
  "rich>=13.0.0",
  "pyyaml>=6.0",
  "google-generativeai>=0.3.0",
  "questionary>=2.0.0",
  "ruamel.yaml>=0.17.0",
  "semver",
  "jinja2",
  "chardet",
  "jsonschema",
]
requires-python = ">= 3.10"

[project.optional-dependencies]
test = [
    "coverage",  # testing
    "pytest",  # testing
    "ruff",  # linting
    "ty", # checking types
    "ipdb", # debugging
]

[project.urls]
bugs = "https://github.com/BigSpoon33/obsidian_capsule_cli/issues"
changelog = "https://github.com/BigSpoon33/obsidian_capsule_cli/blob/master/changelog.md"
homepage = "https://github.com/BigSpoon33/obsidian_capsule_cli"

[project.scripts]
capsule = "capsule.cli:main"
cap = "capsule.cli:main"

[tool.ty]
# All rules are enabled as "error" by default; no need to specify unless overriding.
# Example override: relax a rule for the entire project (uncomment if needed).
# rules.TY015 = "warn"  # For invalid-argument-type, warn instead of error.

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
]

[tool.uv]
package = true
